--worker-close after linking
After library linking is done, additional worker-close action is to be run.
A dummy (libname).dylib.metadata is created from the resultant dylib as input,
so that the worker-close action will be followed by the linking step.
That dummy file is included in HaskellLinkInfo (to be forced).

worker-close action is associated with worker-target-id

Change-Id: I82c507378472fe7e7b55bc7a67e929d193af7580