Haskell non-incremental: use -dynamic-too as well
This is assumed to exist because of usage in _modules_by_name which
declares an output for the dyn_o on static link style.

According to Ian, we probably should remove -dynamic-too in the future
because of incremental builds obviating the benefit. For ghc --make
mode, it still has a benefit though.

Change-Id: I3ba9b5e435cf1fe1d09870e2fb47bb92875c7347